requirements.txt
fastapi
httpx
pydantic
pytest
pytest-asyncio
python-decouple
python-dotenv
psycopg2
rich
scipy
sqlalchemy
uvicorn